medical_diagnosis_chatbot/requirements/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements/requirements-dev.txt pyproject.toml
#
aiohttp==3.9.3
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   langchain-core
    #   openai
    #   starlette
attrs==23.2.0
    # via aiohttp
black==24.3.0
    # via brag (pyproject.toml)
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   uvicorn
cobble==0.1.3
    # via mammoth
colorama==0.4.6
    # via
    #   click
    #   loguru
    #   tqdm
coloredlogs==15.0.1
    # via onnxruntime
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
distro==1.9.0
    # via openai
fastapi==0.110.0
    # via brag (pyproject.toml)
fastembed==0.2.2
    # via
    #   brag (pyproject.toml)
    #   qdrant-client
filelock==3.13.3
    # via huggingface-hub
flatbuffers==24.3.7
    # via onnxruntime
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.3.1
    # via huggingface-hub
greenlet==3.0.3
    # via sqlalchemy
grpcio==1.62.1
    # via
    #   grpcio-tools
    #   qdrant-client
grpcio-tools==1.62.1
    # via qdrant-client
gunicorn==21.2.0
    # via brag (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.4
    # via httpx
httpx[http2]==0.27.0
    # via
    #   openai
    #   qdrant-client
huggingface-hub==0.20.3
    # via
    #   fastembed
    #   tokenizers
humanfriendly==10.0
    # via coloredlogs
hyperframe==6.0.1
    # via h2
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.1.13
    # via
    #   brag (pyproject.toml)
    #   langchain-experimental
langchain-community==0.0.29
    # via
    #   brag (pyproject.toml)
    #   langchain
langchain-core==0.1.33
    # via
    #   langchain
    #   langchain-community
    #   langchain-experimental
    #   langchain-text-splitters
    #   langgraph
langchain-experimental==0.0.55
    # via brag (pyproject.toml)
langchain-text-splitters==0.0.1
    # via langchain
langgraph==0.0.30
    # via brag (pyproject.toml)
langsmith==0.1.31
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
loguru==0.7.2
    # via fastembed
lxml==5.1.0
    # via python-docx
mammoth==1.7.1
    # via brag (pyproject.toml)
marshmallow==3.21.1
    # via dataclasses-json
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
numpy==1.26.4
    # via
    #   brag (pyproject.toml)
    #   fastembed
    #   langchain
    #   langchain-community
    #   onnx
    #   onnxruntime
    #   pandas
    #   qdrant-client
    #   scipy
onnx==1.16.0
    # via fastembed
onnxruntime==1.17.1
    # via fastembed
openai==1.14.3
    # via brag (pyproject.toml)
orjson==3.9.15
    # via langsmith
packaging==23.2
    # via
    #   black
    #   gunicorn
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   onnxruntime
pandas==2.2.2
    # via brag (pyproject.toml)
pathspec==0.12.1
    # via black
platformdirs==4.2.0
    # via black
portalocker==2.8.2
    # via qdrant-client
protobuf==4.25.3
    # via
    #   grpcio-tools
    #   onnx
    #   onnxruntime
psycopg2-binary==2.9.9
    # via brag (pyproject.toml)
pugsql==0.2.4
    # via brag (pyproject.toml)
pydantic==2.6.4
    # via
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.16.3
    # via pydantic
pydantic-settings==2.2.1
    # via brag (pyproject.toml)
pypdf2==3.0.1
    # via brag (pyproject.toml)
pyreadline3==3.4.1
    # via humanfriendly
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.1.0
    # via brag (pyproject.toml)
python-dotenv==1.0.1
    # via pydantic-settings
python-multipart==0.0.9
    # via brag (pyproject.toml)
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
qdrant-client[fastembed]==1.8.0
    # via brag (pyproject.toml)
regex==2023.12.25
    # via tiktoken
requests==2.31.0
    # via
    #   fastembed
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   tiktoken
ruff==0.3.4
    # via brag (pyproject.toml)
scipy==1.13.0
    # via brag (pyproject.toml)
sentry-sdk==1.43.0
    # via brag (pyproject.toml)
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==1.4.52
    # via
    #   langchain
    #   langchain-community
    #   pugsql
starlette==0.36.3
    # via fastapi
sympy==1.12
    # via onnxruntime
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.6.0
    # via brag (pyproject.toml)
tokenizers==0.15.2
    # via fastembed
tqdm==4.66.2
    # via
    #   fastembed
    #   huggingface-hub
    #   openai
typing-extensions==4.10.0
    # via
    #   fastapi
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via
    #   qdrant-client
    #   requests
    #   sentry-sdk
uvicorn==0.29.0
    # via brag (pyproject.toml)
win32-setctime==1.1.0
    # via loguru
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
